Fixed directional lights not coming from the right direction, added WinMain support for release builds, fixed post process passes not working out of the box, fixed debug line renderer not being able to render single frame lines

+++ based/include/based/memory/mutexqueue.h
@@ -57,7 +57,7 @@ namespace based::memory
 			return result;
 		}
 
-		inline uint32_t Size()
+		inline size_t Size()
 		{
 			return mBackingQueue.size();
 		}
